WNBA Arbitrage AI Tool/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0

# Enhanced Arbitrage Dependencies
redis==5.0.1
asyncio-throttle==1.0.2
numpy==1.24.3
scipy==1.11.1
psutil==5.9.5
aiohttp==3.8.5

# Testing Dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
httpx==0.24.1